@@ -1,4 +1,7 @@
-from . import server
+try:
+    from . import server
+except ImportError:
+    from bukuserver import server
 
 
 if __name__ == '__main__':
