@@ -90,7 +90,7 @@ class PrettySafeLoader(
 
 
 PrettySafeLoader.add_constructor(
-    u"tag:yaml.org,2002:python/tuple", PrettySafeLoader.construct_python_tuple
+    "tag:yaml.org,2002:python/tuple", PrettySafeLoader.construct_python_tuple
 )
 
 
@@ -268,7 +268,7 @@ class TestBukuDb(unittest.TestCase):
             for index, tagset in from_db:
                 # checking if new tags added to bookmark
                 self.assertTrue(split_and_test_membership(new_tags, tagset))
-                # checking if old tags still exist for boomark
+                # checking if old tags still exist for bookmark
                 self.assertTrue(split_and_test_membership(old_tagsets[index], tagset))
 
     def test_delete_tag_at_index(self):
@@ -277,7 +277,7 @@ class TestBukuDb(unittest.TestCase):
             self.bdb.add_rec(*bookmark)
 
         get_tags_at_idx = lambda i: self.bdb.get_rec_by_id(i)[3]
-        # list of two-tuples, each containg bookmark index and corresponding tags
+        # list of two-tuples, each containing bookmark index and corresponding tags
         tags_by_index = [
             (i, get_tags_at_idx(i)) for i in inclusive_range(1, len(self.bookmarks))
         ]
@@ -1279,7 +1279,7 @@ def test_load_chrome_database(chrome_db, add_pt):
     res_yaml_file = chrome_db[1] if add_pt else chrome_db[2]
     dump_data = False  # NOTE: change this value to dump data
     if not dump_data:
-        with open(res_yaml_file, "r") as f:
+        with open(res_yaml_file, "r", encoding="utf8", errors="surrogateescape") as f:
             try:
                 res_yaml = yaml.load(f, Loader=yaml.FullLoader)
             except RuntimeError:
@@ -1296,7 +1296,7 @@ def test_load_chrome_database(chrome_db, add_pt):
         assert call_args_list_dict == res_yaml
     # dump data for new test
     if dump_data:
-        with open(res_yaml_file, "w") as f:
+        with open(res_yaml_file, "w", encoding="utf8", errors="surrogateescape") as f:
             yaml.dump(call_args_list_dict, f)
         print("call args list dict dumped to:{}".format(res_yaml_file))
 
@@ -1326,7 +1326,7 @@ def test_load_firefox_database(firefox_db, add_pt):
     dump_data = False  # NOTE: change this value to dump data
     res_yaml_file = firefox_db[1] if add_pt else firefox_db[2]
     if not dump_data:
-        with open(res_yaml_file, "r") as f:
+        with open(res_yaml_file, "r", encoding="utf8", errors="surrogateescape") as f:
             res_yaml = yaml.load(f, Loader=PrettySafeLoader)
     # init
     import buku
@@ -1339,7 +1339,7 @@ def test_load_firefox_database(firefox_db, add_pt):
     if not dump_data:
         assert call_args_list_dict == res_yaml
     if dump_data:
-        with open(res_yaml_file, "w") as f:
+        with open(res_yaml_file, "w", encoding="utf8", errors="surrogateescape") as f:
             yaml.dump(call_args_list_dict, f)
         print("call args list dict dumped to:{}".format(res_yaml_file))
 
@@ -1402,7 +1402,7 @@ def test_exportdb_single_rec(tmpdir):
         db.add_rec("http://example.com")
         exp_file = tmpdir.join("export")
         db.exportdb(exp_file.strpath)
-        with open(exp_file.strpath) as f:
+        with open(exp_file.strpath, encoding="utf8", errors="surrogateescape") as f:
             assert f.read()
 
 
