@@ -5,6 +5,6 @@ def test_bukuserver_requirement(monkeypatch):
     monkeypatch.setattr(setuptools, 'setup', m_setup)
     import setup
     setup_l = setup.server_require
-    with open('bukuserver/requirements.txt') as f:
+    with open('bukuserver/requirements.txt', encoding="utf8", errors="surrogateescape") as f:
         bs_l = [x for x in f.read().splitlines() if not x.startswith('buku')]
     assert bs_l == setup_l
